feat(mookme): add a basis for an extensive filtering strategy system
