+++ packages/mookme/src/loaders/filter-strategies/previous-commit-filter.ts
@@ -1,0 +1,11 @@
+import { PackageHook } from '../../types/hook.types';
+import { FilterStrategy } from './base-filter';
+
+/**
+ * A base class for denoting a strategy used to filter hooks to run
+ */
+export class PreviousCommitFilterStrategy extends FilterStrategy {
+  async filter(hooks: PackageHook[]): Promise<PackageHook[]> {
+    return hooks;
+  }
+}
